refactor(esbabbler): extract shared suggestion list keyboard navigation
